# apps/backend/usersrole/docs/contracts/profile-service.openapi.yaml
openapi: 3.1.0
info:
  title: profile-service
  version: 1.0.0
  summary: Profiles, addresses and icons, frozen against the usersrole Spring controllers.
  description: |
    The profile half of the usersrole split: 15 operations, all under
    `/api/profiles`.

    Frozen against the behaviour of `ProfilesController` and `ProfileService`,
    not against the springdoc document they publish — springdoc reports every
    handler as `200 */*` and mis-types both icon uploads as `application/json`,
    because it reads return types and parameter types rather than what the
    handlers build and consume. Every status and media type below was
    transcribed from the handler bodies and the `GlobalExceptionHandler`.
    `docs/contracts/README.md` lists every deviation from the served document
    and why it is there, and `scripts/check-contract-drift.mjs` fails the build
    when a deviation appears that is not on that list.

    This service authorizes from the verified token minted by `identity-service`
    and makes no call to it on the request path. That is a change from how the
    JVM authorizes today, not a transcription of it: today every request
    re-reads the principal, its grants and its profile from the database. What
    that change costs is recorded in `x-authority-freshness` and
    `x-profile-id-fallback` rather than left to be discovered.
  license:
    name: PolyForm-Noncommercial-1.0.0
    url: https://polyformproject.org/licenses/noncommercial/1.0.0/
servers:
...
security:
  - bearerAuth: []
...
paths:
  /api/profiles:
    get:
      tags: [profiles]
      operationId: getProfiles
      summary: List profiles, one page at a time.
      x-authorization:
        rule: ADMIN
        preAuthorize: "hasAuthority('ADMIN')"
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: A page of profiles as a bare array. No envelope, no total count.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [profiles]
      operationId: createProfile
      summary: Create a profile for a user.
      description: |
        One profile per user. A second create for the same user id is 409.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_BODY_USER_ID
        preAuthorize: "hasAuthority('ADMIN') or #profile.userId() == authentication.principal.getUserId()"
        note: |
          The only rule in either service that reads the request body rather than
          a path variable or a claim. A caller may create a profile for
          themselves, an ADMIN for anyone.
      x-cross-boundary-read:
        today: |
          `ProfileService.createProfile` calls `UserRepository.findById` to prove
          the user exists before inserting, and resolves the acting principal's
          user id by email for the audit columns.
        frozenAs: |
          Neither read survives the split. The acting user id comes from the
          `user_id` claim, and the existence check becomes a foreign key on
          `auth.profiles.user_id` — enforced by Postgres, stronger than the
          application check it replaces, and free on the request path. The 404
          below is preserved: a foreign-key violation on insert is reported as a
          missing user, not as a 500.
      parameters:
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileCreateRequestDTO'
      responses:
        '201':
          description: Profile created. springdoc reports 200; the handler builds `HttpStatus.CREATED`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/profiles/{profileId}:
    get:
      tags: [profiles]
      operationId: getProfileById
      summary: Read one profile.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_PROFILE_ID
        preAuthorize: "hasAuthority('ADMIN') or #profileId == authentication.principal.getProfileId()"
        note: Subject to the `profile_id` claim fallback described in `x-profile-id-fallback`.
      parameters:
        - $ref: '#/components/parameters/ProfileId'
      responses:
        '200':
          description: The profile, with its addresses and icon metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [profiles]
      operationId: updateProfileById
      summary: Replace a profile's name fields and birthdate.
      description: '`userId` is not in the update body: a profile cannot be moved between users.'
      x-authorization:
        rule: ADMIN_OR_SELF_BY_PROFILE_ID
        preAuthorize: "hasAuthority('ADMIN') or #profileId == authentication.principal.getProfileId()"
        note: Subject to the `profile_id` claim fallback described in `x-profile-id-fallback`.
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequestDTO'
      responses:
        '200':
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [profiles]
      operationId: deleteProfileById
      summary: Delete a profile with its addresses and icon.
      description: |
        A no-op for an id that does not exist, and still 204.

        Deleting your own profile changes what your existing token can do, and
        the split changes it differently — see `x-behaviour-change`.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_PROFILE_ID
        preAuthorize: "hasAuthority('ADMIN') or #profileId == authentication.principal.getProfileId()"
      x-behaviour-change:
        change: |
          After deleting their own profile, a non-admin caller reusing the same
          token gets **204** here where today they get 403.
        classification: security
        today: |
          The principal is rehydrated per request, so once the row is gone
          `getProfileId()` returns null, the predicate is false, and this
          operation answers 403.
        frozenAs: |
          Under claim-based authorization the stale non-null claim matches the
          path variable, the predicate passes, and the handler deletes without
          reading anything first — `ProfileRepositoryImpl.deleteById` has no
          existence check, which is also why this operation has no 404 in its
          response set at all. So the caller gets a silent success for a delete
          that matched no row.
        reasoning: |
          This is the wider half of the widening described in
          `x-stale-profile-claim`: the eight operations that look up first
          degrade to 404, which at least says nothing was there, while this one
          and `DELETE /api/profiles/{profileId}/icon` report success. Recorded
          rather than corrected, because correcting it means an existence check
          before authorizing — the per-request read the split removes.
        seeAlso: x-stale-profile-claim
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '204':
          description: |
            Deleted, no body. springdoc reports `200` with a `Profile` body
            because the handler's declared return type is
            `ResponseEntity<Profile>`; the body it builds is `noContent()`.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/profiles/by-user/{userId}:
    get:
      tags: [profiles]
      operationId: getProfileByUserId
      summary: Read a user's profile.
      description: |
        The operation `identity-service` clients use once the `User` payload no
        longer embeds the profile aggregate.
      x-behaviour-change:
        change: 'Moved from `/api/profiles/user/{userId}`.'
        previously: |
          Served one segment under `/api/profiles`, where `user` sits in the
          `{profileId}` slot and the pattern tied with
          `/api/profiles/{profileId}/icon` — making `PUT` and `DELETE
          /api/profiles/user/icon` answer 500. See `x-path-precedence`.
        frozenAs: |
          `/api/profiles/by-user/{userId}`, with nothing left on the old path: an
          alias would be the same pair of patterns and re-create the tie.
        clientImpact: |
          The application moved in the same change, so this is a change to the
          served API rather than a divergence from it. The only caller,
          `libs/frontend/usersui/data-access/src/lib/profiles/profiles.service.ts`,
          moved with it; a client left on the old path reaches the icon routes
          and gets 400 or 406, never this operation.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_USER_ID
        preAuthorize: "hasAuthority('ADMIN') or #userId == authentication.principal.getUserId()"
        note: |
          Keyed on `user_id`, which is always present in the token, so this
          operation is unaffected by the `profile_id` staleness.
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: The profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [profiles]
      operationId: updateProfileByUserId
      summary: Replace a user's profile.
      x-behaviour-change:
        change: 'Moved from `/api/profiles/user/{userId}`.'
        previously: |
          Served one segment under `/api/profiles`, where `user` sits in the
          `{profileId}` slot and the pattern tied with
          `/api/profiles/{profileId}/icon` — making `PUT` and `DELETE
          /api/profiles/user/icon` answer 500. See `x-path-precedence`.
        frozenAs: |
          `/api/profiles/by-user/{userId}`, with nothing left on the old path: an
          alias would be the same pair of patterns and re-create the tie.
        clientImpact: |
          The application moved in the same change, so this is a change to the
          served API rather than a divergence from it. The only caller,
          `libs/frontend/usersui/data-access/src/lib/profiles/profiles.service.ts`,
          moved with it; a client left on the old path reaches the icon routes
          and gets 400 or 406, never this operation.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_USER_ID
        preAuthorize: "hasAuthority('ADMIN') or #userId == authentication.principal.getUserId()"
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequestDTO'
      responses:
        '200':
          description: The updated profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [profiles]
      operationId: deleteProfileByUserId
      summary: Delete a user's profile with its addresses and icon.
      description: A no-op for a user id with no profile, and still 204.
      x-behaviour-change:
        change: 'Moved from `/api/profiles/user/{userId}`.'
        previously: |
          Served one segment under `/api/profiles`, where `user` sits in the
          `{profileId}` slot and the pattern tied with
          `/api/profiles/{profileId}/icon` — making `PUT` and `DELETE
          /api/profiles/user/icon` answer 500. See `x-path-precedence`.
        frozenAs: |
          `/api/profiles/by-user/{userId}`, with nothing left on the old path: an
          alias would be the same pair of patterns and re-create the tie.
        clientImpact: |
          The application moved in the same change, so this is a change to the
          served API rather than a divergence from it. The only caller,
          `libs/frontend/usersui/data-access/src/lib/profiles/profiles.service.ts`,
          moved with it; a client left on the old path reaches the icon routes
          and gets 400 or 406, never this operation.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_USER_ID
        preAuthorize: "hasAuthority('ADMIN') or #userId == authentication.principal.getUserId()"
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '204':
          description: Deleted, no body. springdoc reports 200 with a `Profile` body; the handler builds `noContent()`.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/profiles/{profileId}/address:
    post:
      tags: [addresses]
      operationId: addAddress
      summary: Add an address to a profile.
      description: |
        Returns 200, not 201, unlike every other create in either service. The
        handler builds `ResponseEntity.ok`. Frozen as-is: the asymmetry is
        harmless and changing it would be a client-visible change for nothing.

        The response is the whole profile, not the created address, so a client
        that needs the new address id reads it from `addresses`.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_PROFILE_ID
        preAuthorize: "hasAuthority('ADMIN') or #profileId == authentication.principal.getProfileId()"
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequestDTO'
      responses:
        '200':
          description: The profile with the new address in `addresses`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/profiles/{profileId}/address/{addressId}:
    put:
      tags: [addresses]
      operationId: updateAddress
      summary: Replace one of a profile's addresses.
      description: |
        Scoped to the profile: the handler resolves the address inside that
        profile's own address set, so an `addressId` belonging to another profile
        is 404 rather than an edit.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_PROFILE_ID
        preAuthorize: "hasAuthority('ADMIN') or #profileId == authentication.principal.getProfileId()"
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - $ref: '#/components/parameters/AddressId'
        - $ref: '#/components/parameters/AuthorizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequestDTO'
      responses:
        '200':
          description: The profile with the updated address.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [addresses]
      operationId: deleteAddress
      summary: Delete one of a profile's addresses.
      x-authorization:
        rule: ADMIN_OR_SELF_BY_PROFILE_ID
        preAuthorize: "hasAuthority('ADMIN') or #profileId == authentication.principal.getProfileId()"
      description: |
        Scoped to the profile in the path, and 404 when the address does not
        belong to it. This is a transcription, not a deviation: the scoping is in
        `AddressDaoPostgres.deleteByProfileIdAndAddressId`, whose statement is
        `DELETE FROM auth.addresses WHERE address_id = :addressId AND profile_id
        = :profileId`, and `ProfileService.deleteAddress` raises
        `ResourceNotFoundException` when it deletes no row.

        It was not always so, and the history is worth keeping: until recently
        the handler took `profileId`, used it for the authorization check, and
        then deleted by `addressId` alone, so any authenticated principal with a
        profile could delete any address in the table by guessing a sequential id
        and get 204 for it. This contract specified the scoped behaviour while
        that was still true, because writing an insecure direct object reference
        into the document the Go services are built and parity-tested against
        would have had the generated suite certify it. The application has since
        been fixed independently, so contract and code now agree.
      parameters:
        - $ref: '#/components/parameters/ProfileId'
        - $ref: '#/components/parameters/AddressId'
        - $ref: '#/components/parameters/AuthorizationHeader'
      responses:
        '204':
          description: Deleted, no body. springdoc reports 200 with a `Profile` body; the handler builds `noContent()`.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: |
            The address does not belong to the profile in the path, so the
            scoped delete matched no row. `Address not found with id {addressId}
            for profile with id {profileId}` as `text/plain`. Unreachable for any
            caller acting on their own data.
          content:
            text/plain:
              schema:
                type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        `Authorization: Bearer <jwt>`, minted by `identity-service` and verified
        here against the shared HS256 secret. The claim set is documented as
        `x-jwt-claims` in the identity contract; this service reads `roles`,
        `user_id` and `profile_id` from it and calls nothing.
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: |
        Zero-based page index. Out-of-range input is clamped, not rejected: a
        negative page is served as page 0 with a 200, never a 400.
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
    Size:
      name: size
      in: query
      required: false
      description: |
        Page size, bounded server-side to at most 500 rows so the listing always
        issues a bounded query whatever the caller asks for. Like `page`, out-of
        -range input is clamped rather than rejected: `size=0` is served as 1 and
        `size=100000` as 500, both with a 200. `minimum` and `maximum` here
        describe the range the service will actually serve, not a range it
        validates — a request outside it succeeds.
      schema:
        type: integer
        format: int32
        default: 100
        minimum: 1
        maximum: 500
    ProfileId:
      name: profileId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    AuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: |
        Declared explicitly because the handler reads the raw header to resolve
        the acting user for the audit columns. Redundant with the `bearerAuth`
        security scheme and kept only because removing it would change the served
        document. In the Go service the acting user comes from the verified
        `user_id` claim and this parameter disappears.
      schema:
        type: string
  responses:
    BadRequest:
      description: |
        Bean-validation failures produce a JSON object of field name to message;
        an unparseable body produces the fixed `text/plain` string
        `Request body is invalid. Please check the format and try again.`
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrors'
        text/plain:
          schema:
            type: string
    Unauthorized:
      description: |
        No token, an expired token, or a token that fails verification. Written
        through `HttpServletResponse.sendError`, so the body is the servlet
        container's error representation. Clients key off the status and the
        header, never the body.
      headers:
        Access-Denied-Reason:
          description: '`Authentication Required`'
          required: true
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ContainerError'
    Forbidden:
      description: |
        Authenticated but not permitted by the operation's `x-authorization`
        rule. Written through `sendError`, so the body is the container's error
        representation.
      headers:
        Access-Denied-Reason:
          description: '`Not Authorized`'
          required: true
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ContainerError'
    NotFound:
      description: |
        `ResourceNotFoundException`. The body is the exception message as
        `text/plain`, e.g. `Profile not found with id 42`.
      content:
        text/plain:
          schema:
            type: string
...
  schemas:
    ProfileCreateRequestDTO:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        middleName:
          type: string
        lastName:
          type: string
          minLength: 1
        birthdate:
          type: string
          format: date-time
        userId:
          type: integer
          format: int64
          description: The user this profile belongs to. Immutable once created.
      required: [firstName, lastName, birthdate, userId]
...
    Profile:
      type: object
      description: |
        The profile aggregate. Returned whole by every operation on this service
        except the icon download and the deletes, including the address and icon
        writes, which answer with the parent profile rather than the subresource.
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
        middleName:
          type: [string, 'null']
        lastName:
          type: string
        birthdate:
          type: string
          format: date-time
        userId:
          type: integer
          format: int64
        addresses:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Address'
        icon:
          description: |
            The icon's metadata and bytes, or null when the profile has none.
            Base64 in JSON, so a profile read carries the whole image; the
            dedicated `image/png` download exists for clients that want the bytes
            without the envelope.
          oneOf:
            - $ref: '#/components/schemas/ProfileIcon'
            - type: 'null'
        createdByUserId:
          type: integer
          format: int64
        createdTime:
          type: string
          format: date-time
        modifiedByUserId:
          type: integer
          format: int64
        modifiedTime:
          type: string
          format: date-time
...